requirements.txt
pandas
lxml
sqlalchemy
pymysql
python-dotenv
tabulate